hex/decimal
We currently have a mix of hexadecimal and decimal used in different places.  I am trying to get a degree of consistency in using decimal for memory addresses at least.  I think the longer term aim should be to switch to decimal for instructions too for consistency (these are better displayed as hex, but perhaps even better not seen at all!  This latter is incorrect though as it is good for students to see the instructions as 'data')